+++ areas/storage/src/AzureMcp.Storage/Commands/StorageJsonContext.cs
@@ -16,6 +16,7 @@ namespace AzureMcp.Storage.Commands;
 
 [JsonSerializable(typeof(BlobListCommand.BlobListCommandResult))]
 [JsonSerializable(typeof(BlobDetailsCommand.BlobDetailsCommandResult))]
+[JsonSerializable(typeof(BlobUploadCommand.BlobUploadCommandResult))]
 [JsonSerializable(typeof(BatchSetTierCommand.BatchSetTierCommandResult))]
 [JsonSerializable(typeof(AccountListCommand.AccountListCommandResult), TypeInfoPropertyName = "AccountListCommandResult")]
 [JsonSerializable(typeof(AccountCreateCommand.AccountCreateCommandResult), TypeInfoPropertyName = "AccountCreateCommandResult")]